improve: communication history back to as it is and change in automated communication history
